+++ goodq_version.py
@@ -1,4 +1,4 @@
 """Canonical GoodQ4All software version."""
 
-GOODQ_VERSION = "2.5.8-rc1"
+GOODQ_VERSION = "2.5.8-rc2"
 __version__ = GOODQ_VERSION